Avoid cost of binding flush
Flushing cached BindingReferences takes a few mS, which is Ok as a one off
but adds significant cost when adding many bindings. If we instead keep a
cache generation counter, we can use a fast check as to whether a binding
is still valid before using it and rebind if not.