Fix bug in renaming namespace classes imported into other classes
Renaming namespaces did not update the imports of referencing classes
correctly - the old name was left behind in the import list.
Also add change object for updating class comments. Not used by the
ClassCommentPlugin yet though.
Fix bug in RenameClassRefactoring where the wrong name was used to
determine whether renaming a new class in the model.